Create useSignUpForm hook to separate logic from UI
